Move ``_filter_stmts`` into its own file
